Do a slight refactoring
